+++ src/main/scala/aoc/Day07.scala
@@ -3,13 +3,13 @@ package aoc
 import aoc.util.{Intcode, NoInput}
 
 object Day07 {
-  def first(programStr: String): Int = {
-    val program = programStr.split(',').map(_.toInt)
+  def first(programStr: String): Long = {
+    val program = programStr.split(',').map(_.toLong)
     (0 to 4).permutations.map(perm => emulate(program, perm)).max
   }
 
-  def emulate(program: Array[Int], phases: Seq[Int]): Int = {
-    var signal = 0
+  def emulate(program: Array[Long], phases: Seq[Int]): Long = {
+    var signal = 0L
     for (phase <- phases) {
       val amplifier = new Intcode(program.clone)
       amplifier.execute(Seq(phase, signal).iterator)
@@ -18,13 +18,13 @@ object Day07 {
     signal
   }
 
-  def second(programStr: String): Int = {
-    val program = programStr.split(',').map(_.toInt)
+  def second(programStr: String): Long = {
+    val program = programStr.split(',').map(_.toLong)
     (5 to 9).permutations.map(perm => emulateWithFeedback(program, perm)).max
   }
 
-  def emulateWithFeedback(program: Array[Int], phases: Seq[Int]): Int = {
-    def executeOnce(amplifier: Intcode, input: Seq[Int]): Unit = {
+  def emulateWithFeedback(program: Array[Long], phases: Seq[Int]): Long = {
+    def executeOnce(amplifier: Intcode, input: Seq[Long]): Unit = {
       try {
         amplifier.execute(input.iterator, reset = false)
       } catch {
@@ -35,7 +35,7 @@ object Day07 {
     val amplifiers = phases.map(_ => new Intcode(program.clone))
     (amplifiers zip phases).foreach(p => executeOnce(p._1, Seq(p._2)))
 
-    var signal = 0
+    var signal = 0L
     while (!amplifiers.last.halted) {
       for (amp <- amplifiers) {
         executeOnce(amp, Seq(signal))
